pyproject.toml
[tool.poetry]
name = "job-spider"
version = "0.1.0"
description = ""
authors = ["jie <yanjya2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.1"
requests = "^2.31.0"
ipykernel = "^6.29.4"
langchain = "^0.1.16"
streamlit = "^1.33.0"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.3"
pypdf = "^4.2.0"
chroma = "^0.2.0"
langchain-community = "^0.0.32"
langchain-experimental = "^0.0.57"
tabulate = "^0.9.0"
chromadb = "^0.4.24"
wordcloud = "^1.9.3"
matplotlib = "^3.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
